+++ src/main/resources/web-driver.properties
@@ -1,4 +1,6 @@
-webDriver-executable=chromedriver.exe
+webDriver-executable=/usr/bin/chromedriver
+chrome-executable=/opt/google/chrome/chrome
+chrome-version=117.0.5938.88
 user-agent=Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/92.0.4515.131 Safari/537.36
 page-load-timeout=15000
 implicitly-wait=5000
